+++ lib/components/base-components/NormalComponent/NormalComponent.ts
@@ -46,7 +46,6 @@ import { parsePinNumberFromLabelsOrThrow } from "lib/utils/schematic/parsePinNum
 import { getNumericSchPinStyle } from "lib/utils/schematic/getNumericSchPinStyle"
 import type { INormalComponent } from "./INormalComponent"
 import { Trace } from "lib/components/primitive-components/Trace/Trace"
-import { TraceHint } from "lib/components/primitive-components/TraceHint"
 
 const debug = Debug("tscircuit:core")
 
@@ -1196,563 +1195,19 @@ export class NormalComponent<
         return
       }
 
-      // Check if this component should use hub-based MSP routing (fallback for non-subcircuit contexts)
-      const hubMSPTraces = this._tryGenerateHubMSPTraces()
-      if (hubMSPTraces.length > 0) {
-        console.log(`🔗 ${this.getString()} using hub MSP routing with ${hubMSPTraces.length} traces`)
-        for (const trace of hubMSPTraces) {
-          this.add(trace)
-        }
-        // Log the complete MSP chain being built
-        console.log(`⛓️  MSP Chain building: ${this.getString()} contributes to minimum spanning tree`)
-        return // Skip normal routing if hub MSP was applied
-      }
-
+      // Use standard trace creation for all other cases
       for (const [pinName, target] of Object.entries(props.connections)) {
         const targets = Array.isArray(target) ? target : [target]
-        const pinSelector = `${this.getSubcircuitSelector()} > port.${pinName}`
-
-        // Use MSP routing for eligible connections
-        if (this._shouldUseMSPRouting(pinName, targets)) {
-          const mspTraces = this._generateMSPTraces(pinSelector, targets)
-          for (const trace of mspTraces) {
-            this.add(trace)
-          }
-        } else {
-          // Use original star pattern for other connections
-          for (const targetPath of targets) {
-            this.add(
-              new Trace({
-                from: pinSelector,
-                to: targetPath as string,
-              }),
-            )
-          }
+        for (const targetPath of targets) {
+          this.add(
+            new Trace({
+              from: `${this.getSubcircuitSelector()} > port.${pinName}`,
+              to: targetPath as string,
+            }),
+          )
         }
       }
     }
   }
 
-  /**
-   * Attempts to generate hub-based MSP routing when multiple components connect to the same target.
-   * This handles cases like R2→R1, R3→R1, R4→R1 by creating a chain R4→R3→R2→R1.
-   */
-  _tryGenerateHubMSPTraces(): any[] {
-    const { _parsedProps: props } = this
-    if (!props.connections) return []
-
-    // Look for single-target component connections (potential hub pattern)
-    const singleComponentConnections = Object.entries(props.connections)
-      .filter(([, target]) => {
-        const targets = Array.isArray(target) ? target : [target]
-        return targets.length === 1 && /^[A-Z]+\d+\.[A-Z0-9]+$/i.test(targets[0])
-      })
-
-    console.log(`🎯 ${this.getString()} found ${singleComponentConnections.length} single component connections:`, singleComponentConnections)
-    if (singleComponentConnections.length === 0) return []
-
-    // For each connection, check if we should participate in hub MSP
-    const traces: any[] = []
-    for (const [pinName, target] of singleComponentConnections) {
-      const targetPin = Array.isArray(target) ? target[0] : target
-      console.log(`🔍 ${this.getString()} analyzing connection ${pinName} → ${targetPin}`)
-      const hubMSPTrace = this._generateHubMSPTrace(pinName, targetPin)
-      if (hubMSPTrace) {
-        console.log(`✅ ${this.getString()} generated hub MSP trace!`)
-        traces.push(hubMSPTrace)
-      } else {
-        console.log(`❌ ${this.getString()} no hub MSP trace (closest to hub)`)
-      }
-    }
-
-    return traces
-  }
-
-  /**
-   * Generates a single hub MSP trace by finding the optimal path.
-   * Uses a coordinated approach: one component connects to hub, others form chain.
-   */
-  _generateHubMSPTrace(pinName: string, targetPin: string): any | null {
-    const subcircuit = this.getSubcircuit()
-    if (!subcircuit) return null
-
-    const thisPosition = this._getComponentPosition(`${this._parsedProps.name}.${pinName}`)
-    if (!thisPosition) return null
-
-    // Get the target hub position
-    const targetPosition = this._getComponentPosition(targetPin)
-    if (!targetPosition) return null
-
-    console.log(`📍 ${this.getString()} positions - self: (${thisPosition.x}, ${thisPosition.y}), hub: (${targetPosition.x}, ${targetPosition.y})`)
-
-    // Find all components targeting the same hub (including self)
-    const allComponents = subcircuit.selectAll('resistor, capacitor, inductor')
-    const hubComponents: Array<{name: string, component: any, position: {x: number, y: number}, distanceToHub: number}> = []
-    
-    console.log(`🔍 ${this.getString()} scanning ${allComponents.length} components for hub pattern...`)
-    for (const component of allComponents) {
-      try {
-        const componentProps = component._parsedProps
-        if (componentProps?.connections) {
-          for (const [compPinName, compTarget] of Object.entries(componentProps.connections)) {
-            const compTargets = Array.isArray(compTarget) ? compTarget : [compTarget]
-            if (compTargets.includes(targetPin)) {
-              const componentPosition = this._getComponentPosition(`${componentProps.name}.${compPinName}`)
-              if (componentPosition) {
-                const distanceToHub = Math.sqrt(
-                  Math.pow(targetPosition.x - componentPosition.x, 2) + 
-                  Math.pow(targetPosition.y - componentPosition.y, 2)
-                )
-                console.log(`📏 Found hub component: ${componentProps.name} at (${componentPosition.x}, ${componentPosition.y}), distance to hub: ${distanceToHub.toFixed(3)}`)
-                hubComponents.push({
-                  name: componentProps.name,
-                  component,
-                  position: componentPosition,
-                  distanceToHub
-                })
-              }
-            }
-          }
-        }
-      } catch (error) {
-        continue
-      }
-    }
-
-    if (hubComponents.length <= 1) {
-      console.log(`❌ ${this.getString()} no hub pattern detected (only ${hubComponents.length} components)`)
-      return null
-    }
-
-    // Sort by distance to hub - closest connects directly to hub
-    hubComponents.sort((a, b) => a.distanceToHub - b.distanceToHub)
-    const closestToHub = hubComponents[0]
-    
-    console.log(`🎯 ${this.getString()} hub routing order:`, hubComponents.map((hc, i) => 
-      `${i}: ${hc.name} (dist: ${hc.distanceToHub.toFixed(3)})`
-    ).join(', '))
-    
-    // Find this component in the list
-    const thisComponent = hubComponents.find(hc => hc.name === this._parsedProps.name)
-    if (!thisComponent) return null
-
-    // If this is the closest component to hub, let it use normal routing to hub
-    if (thisComponent === closestToHub) {
-      console.log(`🏆 ${this.getString()} is closest to hub → using normal routing`)
-      return null
-    }
-
-    // Otherwise, connect to the previous component in the sorted chain
-    const thisIndex = hubComponents.indexOf(thisComponent)
-    const targetComponent = hubComponents[thisIndex - 1]
-    
-    console.log(`🔗 ${this.getString()} chain position ${thisIndex} → connecting to ${targetComponent.name} (position ${thisIndex - 1})`)
-
-    // Create trace to the previous component in chain with debug styling
-    const fromSelector = `${this.getSubcircuitSelector()} > port.${pinName}`
-    const toSelector = `${targetComponent.component.getSubcircuitSelector()} > port.pin1`
-    
-    const alignmentY = (thisPosition.y + targetComponent.position.y) / 2 - 0.5
-    
-    // Debug styling info for visual identification
-    const debugColors = ['#ff0000', '#00ff00', '#0000ff', '#ff00ff', '#ffff00', '#00ffff']
-    const debugDashes = ['5,5', '10,3', '15,5,5,5', '20,10', '8,2,2,2', '12,6,2,6']
-    const color = debugColors[thisIndex % debugColors.length]
-    const dashArray = debugDashes[thisIndex % debugDashes.length]
-    
-    console.log(`🎨 ${this.getString()} debug trace style: color=${color}, dash=${dashArray}, width=${3 + thisIndex}`)
-    
-    const trace = new Trace({
-      from: fromSelector,
-      to: toSelector,
-      schematicRouteHints: [
-        { x: targetComponent.position.x, y: alignmentY }
-      ]
-    })
-    
-    // Set debug properties directly on the trace instance
-    // This is a workaround since the props system isn't working properly
-    trace._debugColor = color
-    trace._debugLineStyle = dashArray === '5,5' ? 'dashed' : 
-                           dashArray === '10,3' ? 'dashed' :
-                           dashArray === '15,5,5,5' ? 'dashed' :
-                           dashArray === '2,2' ? 'dotted' : 'solid'
-    trace._debugWidthMultiplier = 1 + thisIndex * 0.3 // Progressively thicker
-    trace._debugTransparency = 0.8 + thisIndex * 0.05 // Progressively more opaque
-    
-    return trace
-  }
-
-  /**
-   * Determines if MSP routing should be used for the given connection
-   */
-  _shouldUseMSPRouting(pinName: string, targets: string[]): boolean {
-    // Enable MSP for any connection with multiple component targets (not just capacitors)
-    const componentTargets = targets.filter(
-      (target) =>
-        /^[A-Z]+\d+\.[A-Z0-9]+$/i.test(target) && !target.includes("net."),
-    )
-    const hasMultipleComponentTargets = componentTargets.length >= 3
-
-    // Also enable for power/ground connections even with fewer targets
-    const isPowerPin = /VIN|VCC|VDD|VOUT/i.test(pinName)
-    const isGroundPin = /GND|VSS/i.test(pinName)
-    const isPowerConnection =
-      (isPowerPin || isGroundPin) && componentTargets.length >= 2
-      
-    // TODO: Enable MSP for single component target connections (potential hub connections)
-    // This would allow components connecting to the same hub to use coordinated MSP routing
-    // const isSingleComponentConnection = componentTargets.length === 1 && targets.length === 1
-    
-    return hasMultipleComponentTargets || isPowerConnection
-  }
-
-  /**
-   * Generates MSP-optimized traces from connections array
-   */
-  _generateMSPTraces(pinSelector: string, targets: string[]): any[] {
-    // Separate component pins from nets (works with any component type: C, R, L, JP, etc.)
-    const componentPins = targets.filter(
-      (target) =>
-        /^[A-Z]+\d+\.[A-Z0-9]+$/i.test(target) && !target.includes("net."),
-    )
-    const netTargets = targets.filter((target) => target.includes("net."))
-
-    if (componentPins.length >= 1) {
-      const traces: any[] = []
-
-      // Generate optimal MSP (Minimum Spanning Path) routing sequence for any components
-      const optimalSequence = this._computeMSPSequence(
-        componentPins,
-        pinSelector,
-      )
-
-      // Convert pin selectors to full component selectors that can be resolved
-      const convertToFullSelector = (pinSelector: string): string => {
-        // Extract component name (e.g., "C2.1" -> "C2")
-        const [componentName, pinName] = pinSelector.split(".")
-
-        // Get the parent group selector, but find the correct scope for the components
-        // The components are in the same parent group as this chip, not inside this chip
-        const currentSelector = this.getSubcircuitSelector()
-        let parentGroupSelector = ""
-
-        if (currentSelector && currentSelector !== "board") {
-          // Extract the parent group from the current selector
-          // e.g., "group > chip.U1" -> "group"
-          const parts = currentSelector.split(" > ")
-          if (parts.length > 1) {
-            parentGroupSelector = parts.slice(0, -1).join(" > ")
-          } else {
-            parentGroupSelector = parts[0]
-          }
-          return `${parentGroupSelector} > .${componentName} > port.${pinName}`
-        } else {
-          return `.${componentName} > port.${pinName}`
-        }
-      }
-
-      // Create MSP tree chain with consistent routing - horizontal bus approach
-      // Calculate optimal Y-level based on component positions for uniform vertical drops
-      const componentPositions = optimalSequence.map(pinSelector => 
-        this._getComponentPosition(pinSelector)
-      ).filter(pos => pos !== null)
-      
-      // Calculate average Y position of components and offset slightly below for clean routing
-      let alignmentY = -1.25 // Default fallback
-      if (componentPositions.length > 0) {
-        const avgY = componentPositions.reduce((sum, pos) => sum + pos!.y, 0) / componentPositions.length
-        // Position horizontal bus 0.75 units below the average component Y position
-        alignmentY = avgY - 0.75
-      }
-
-      for (let i = 0; i < optimalSequence.length - 1; i++) {
-        const fromSelector = convertToFullSelector(optimalSequence[i])
-        const toSelector = convertToFullSelector(optimalSequence[i + 1])
-        
-        // Get component positions for more precise routing hints
-        const fromPos = this._getComponentPosition(optimalSequence[i])
-        const toPos = this._getComponentPosition(optimalSequence[i + 1])
-        
-        let routeHints: Array<{ x: number; y: number }> = [
-          { x: 0, y: alignmentY }, // Horizontal bus level
-          { x: 1, y: alignmentY }, // Maintain horizontal bus
-        ]
-        
-        // Add specific drop points if we have position info
-        if (fromPos && toPos) {
-          routeHints = [
-            { x: fromPos.x, y: alignmentY }, // Drop from component to bus
-            { x: toPos.x, y: alignmentY },   // Rise from bus to component
-          ]
-        }
-
-        const trace = new Trace({
-          from: fromSelector,
-          to: toSelector,
-          schematicRouteHints: routeHints,
-        })
-        traces.push(trace)
-      }
-
-      // Connect the pin to the first component in the sequence (closest to IC)
-      const firstComponentSelector = convertToFullSelector(optimalSequence[0])
-      const firstComponentPos = this._getComponentPosition(optimalSequence[0])
-      
-      let initialRouteHints: Array<{ x: number; y: number }> = [
-        { x: 0, y: alignmentY }, // Force consistent Y-level
-        { x: 1, y: alignmentY }, // Midpoint guidance
-        { x: 2, y: alignmentY }, // Endpoint guidance
-      ]
-      
-      // Use specific routing if we have position info
-      if (firstComponentPos) {
-        initialRouteHints = [
-          { x: firstComponentPos.x, y: alignmentY }, // Drop to horizontal bus level
-        ]
-      }
-      
-      const initialTrace = new Trace({
-        from: pinSelector,
-        to: firstComponentSelector,
-        schematicRouteHints: initialRouteHints,
-      })
-      traces.push(initialTrace)
-
-      // Connect nets directly to the pin (power distribution)
-      for (const netTarget of netTargets) {
-        traces.push(
-          new Trace({
-            from: pinSelector,
-            to: netTarget,
-          }),
-        )
-      }
-
-      return traces
-    }
-
-    // Fallback to star pattern if not enough components for MSP
-    return targets.map(
-      (target) =>
-        new Trace({
-          from: pinSelector,
-          to: target,
-        }),
-    )
-  }
-
-  /**
-   * Computes optimal MSP (Minimum Spanning Path) sequence for any component routing
-   * Works with capacitors, resistors, jumpers, inductors, or any component type
-   */
-  _computeMSPSequence(
-    componentPins: string[],
-    icPinSelector: string,
-  ): string[] {
-    if (componentPins.length === 0) return []
-    if (componentPins.length === 1) return componentPins
-
-    // Extract component information for distance-based routing
-    // Pattern matches: C1.1, R5.2, JP3.A, L2.1, IC4.VCC, etc.
-    const components = componentPins.map((pin) => {
-      const componentMatch = pin.match(/([A-Z]+)(\d+)\.([A-Z0-9]+)/i)
-      return {
-        pin,
-        prefix: componentMatch?.[1] || "",
-        number: parseInt(componentMatch?.[2] || "0"),
-        pinRef: componentMatch?.[3] || "",
-      }
-    })
-
-    // Group components by type (C, R, L, JP, IC, etc.)
-    const componentsByType = components.reduce(
-      (acc, comp) => {
-        if (!acc[comp.prefix]) acc[comp.prefix] = []
-        acc[comp.prefix].push(comp)
-        return acc
-      },
-      {} as Record<string, typeof components>,
-    )
-
-    // For each component type, compute optimal routing sequence
-    let optimalSequence: string[] = []
-
-    for (const [, comps] of Object.entries(componentsByType)) {
-      if (comps.length < 2) {
-        // Single component, add directly
-        optimalSequence.push(...comps.map((c) => c.pin))
-        continue
-      }
-
-      // Sort by component number for consistent ordering
-      comps.sort((a, b) => a.number - b.number)
-
-      // Apply MSP algorithm: create minimal spanning tree
-      // Strategy: Start from one end, find shortest path through all nodes
-      const mspPath = this._computeMinimalSpanningPath(comps, icPinSelector)
-      optimalSequence.push(...mspPath.map((c) => c.pin))
-    }
-
-    return optimalSequence
-  }
-
-  /**
-   * Computes minimal spanning path using actual physical distance between components
-   * Works with any component type - creates tree-like structure that minimizes routing length
-   * Starts from the component closest to the IC and chains outward
-   */
-  _computeMinimalSpanningPath(
-    components: Array<{
-      pin: string
-      number: number
-      pinRef: string
-      prefix: string
-    }>,
-    icPinSelector: string,
-  ): Array<{ pin: string; number: number; pinRef: string; prefix: string }> {
-    if (components.length <= 2) return components
-
-    // Get physical positions of all components
-    const componentsWithPositions = components
-      .map((comp) => {
-        const position = this._getComponentPosition(comp.pin)
-        return { ...comp, position }
-      })
-      .filter((comp) => comp.position !== null)
-
-    if (componentsWithPositions.length !== components.length) {
-      // Fallback to number-based ordering if positions unavailable
-      return components.sort((a, b) => a.number - b.number)
-    }
-
-    // Get the actual IC position from the pin selector
-    const icComponentName =
-      icPinSelector.split(" > port.")[0].split(" > ").pop() ||
-      icPinSelector.split(".")[0]
-    const icComponent = this.getSubcircuit().selectOne(icComponentName)
-    let icPosition = { x: 0, y: 0 } // Default fallback
-
-    // Try to get IC position from various sources
-    if (
-      icComponent &&
-      "_parsedProps" in icComponent &&
-      icComponent._parsedProps
-    ) {
-      const props = icComponent._parsedProps as any
-      if (props.schX !== undefined && props.schY !== undefined) {
-        icPosition = { x: props.schX, y: props.schY }
-      }
-    }
-
-    // If no explicit position found, estimate based on component layout
-    // ICs are typically positioned to the right of passive components
-    if (icPosition.x === 0 && icPosition.y === 0) {
-      // Find the rightmost component position and place IC to the right of it
-      const maxX = Math.max(
-        ...componentsWithPositions.map((c) => c.position?.x || 0),
-      )
-      icPosition = { x: maxX + 2, y: 0 } // Place IC 2 units to the right of rightmost component
-    }
-
-    // Find the component closest to the IC by actual distance calculation
-
-    // Start with the component closest to the IC
-    let startComponent = componentsWithPositions[0]
-    let minDistanceToIC = Math.sqrt(
-      Math.pow(icPosition.x - (startComponent.position?.x || 0), 2) +
-        Math.pow(icPosition.y - (startComponent.position?.y || 0), 2),
-    )
-
-    for (const comp of componentsWithPositions) {
-      const compPos = comp.position!
-      const distanceToIC = Math.sqrt(
-        Math.pow(icPosition.x - compPos.x, 2) +
-          Math.pow(icPosition.y - compPos.y, 2),
-      )
-
-      if (distanceToIC < minDistanceToIC) {
-        minDistanceToIC = distanceToIC
-        startComponent = comp
-      }
-    }
-
-    const unvisited = componentsWithPositions.filter(
-      (c) => c !== startComponent,
-    )
-    const path = [startComponent]
-
-    // Greedy nearest neighbor approach using actual physical distance
-    while (unvisited.length > 0) {
-      const current = path[path.length - 1]
-      const currentPos = current.position!
-
-      // Find nearest unvisited component by Euclidean distance
-      let nearest = unvisited[0]
-      let nearestPos = nearest.position!
-      let minDistance = Math.sqrt(
-        Math.pow(currentPos.x - nearestPos.x, 2) +
-          Math.pow(currentPos.y - nearestPos.y, 2),
-      )
-
-      for (const comp of unvisited) {
-        const compPos = comp.position!
-        const distance = Math.sqrt(
-          Math.pow(currentPos.x - compPos.x, 2) +
-            Math.pow(currentPos.y - compPos.y, 2),
-        )
-        if (distance < minDistance) {
-          minDistance = distance
-          nearest = comp
-        }
-      }
-
-      path.push(nearest)
-      unvisited.splice(unvisited.indexOf(nearest), 1)
-    }
-
-    return path
-  }
-
-  /**
-   * Gets the physical position of a component by its pin selector
-   */
-  _getComponentPosition(pinSelector: string): { x: number; y: number } | null {
-    try {
-      // Extract component name from pin selector (e.g., "C1.1" -> "C1")
-      const componentName = pinSelector.split(".")[0]
-
-      // Find the component in the circuit using CSS-like selector
-      const component = this.getSubcircuit().selectOne(`.${componentName}`)
-      if (!component) {
-        return null
-      }
-
-      // Try to get the position from various properties
-      if ("_parsedProps" in component && component._parsedProps) {
-        const props = component._parsedProps as any
-
-        if (props.pcbX !== undefined && props.pcbY !== undefined) {
-          return { x: props.pcbX, y: props.pcbY }
-        }
-        if (props.schX !== undefined) {
-          const y = props.schY !== undefined ? props.schY : 0
-          return { x: props.schX, y: y }
-        }
-      }
-
-      // Try to get position from component center after layout
-      if (
-        "_getSchematicCenterPosition" in component &&
-        typeof component._getSchematicCenterPosition === "function"
-      ) {
-        const centerPos = component._getSchematicCenterPosition()
-        if (centerPos) return { x: centerPos.x, y: centerPos.y }
-      }
-
-      return null
-    } catch (error) {
-      return null
-    }
-  }
 }
