+++ lib/components/primitive-components/Group/Group.ts
@@ -1001,8 +1001,9 @@ export class Group<Props extends z.ZodType<any, any, any> = typeof groupProps>
    * before individual components create their traces
    */
   doInitialCreateNetsFromProps() {
-    // Run MSP routing coordination early, before child components create their traces
-    this.doInitialMSPRouting()
+    // Import and run MSP routing coordination early
+    const { coordinateSubcircuitMSPRouting } = require('lib/utils/schematic/MSPRouting')
+    coordinateSubcircuitMSPRouting(this)
   }
 
   doInitialSchematicReplaceNetLabelsWithSymbols() {
@@ -1030,256 +1031,6 @@ export class Group<Props extends z.ZodType<any, any, any> = typeof groupProps>
           symbol_name: "vcc_up",
         })
       }
-    }
-  }
-
-  /**
-   * Coordinate MSP (Minimum Spanning Pathfinding) routing at the subcircuit level
-   * This detects hub patterns and creates coordinated MSP traces to avoid repetition
-   */
-  doInitialMSPRouting() {
-    if (this.root?.schematicDisabled) return
-    if (!this.isSubcircuit) return
-    
-    console.log(`🏭 ${this.getString()} starting subcircuit-level MSP routing coordination`)
-    
-    // Detect all hub patterns in this subcircuit
-    const hubPatterns = this._detectHubPatterns()
-    
-    // Create coordinated MSP traces for each hub pattern
-    for (const hubPattern of hubPatterns) {
-      this._createMSPTracesForHub(hubPattern)
-    }
-    
-    console.log(`🏭 ${this.getString()} completed MSP routing for ${hubPatterns.length} hub patterns`)
-  }
-
-  /**
-   * Detect hub patterns in this subcircuit where multiple components connect to the same target
-   */
-  _detectHubPatterns(): Array<{
-    hubTarget: string
-    hubPosition: { x: number, y: number }
-    components: Array<{
-      name: string
-      component: any
-      position: { x: number, y: number }
-      distanceToHub: number
-      pinName: string
-    }>
-  }> {
-    const allComponents = this.selectAll('resistor, capacitor, inductor')
-    const hubMap = new Map<string, Array<{
-      name: string
-      component: any
-      position: { x: number, y: number }
-      pinName: string
-    }>>()
-    
-    console.log(`🔍 ${this.getString()} scanning ${allComponents.length} components for hub patterns...`)
-    
-    // Group components by their connection targets
-    for (const component of allComponents) {
-      try {
-        const componentProps = component._parsedProps
-        if (componentProps?.connections) {
-          for (const [compPinName, compTarget] of Object.entries(componentProps.connections)) {
-            if (typeof compTarget === 'string' && compTarget.includes('.')) {
-              const targetKey = compTarget
-              if (!hubMap.has(targetKey)) {
-                hubMap.set(targetKey, [])
-              }
-              
-              const position = component._getGlobalSchematicPositionBeforeLayout()
-              hubMap.get(targetKey)!.push({
-                name: componentProps.name,
-                component,
-                position,
-                pinName: compPinName
-              })
-            }
-          }
-        }
-      } catch (error) {
-        console.warn(`⚠️  Error processing component ${component.getString()}:`, error)
-      }
-    }
-    
-    // Filter to only hub patterns (2+ components connecting to same target)
-    const hubPatterns: Array<{
-      hubTarget: string
-      hubPosition: { x: number, y: number }
-      components: Array<{
-        name: string
-        component: any
-        position: { x: number, y: number }
-        distanceToHub: number
-        pinName: string
-      }>
-    }> = []
-    
-    for (const [hubTarget, components] of hubMap.entries()) {
-      if (components.length >= 2) {
-        // Find the hub position (target component position)
-        const hubPosition = this._getComponentPosition(hubTarget)
-        if (!hubPosition) continue
-        
-        // Calculate distances and sort
-        const componentsWithDistances = components.map(comp => ({
-          ...comp,
-          distanceToHub: Math.sqrt(
-            Math.pow(comp.position.x - hubPosition.x, 2) + 
-            Math.pow(comp.position.y - hubPosition.y, 2)
-          )
-        })).sort((a, b) => a.distanceToHub - b.distanceToHub)
-        
-        hubPatterns.push({
-          hubTarget,
-          hubPosition,
-          components: componentsWithDistances
-        })
-        
-        console.log(`🎯 Found hub pattern: ${componentsWithDistances.length} components → ${hubTarget}`)
-        componentsWithDistances.forEach((comp, i) => {
-          console.log(`   ${i}: ${comp.name} at (${comp.position.x}, ${comp.position.y}), distance: ${comp.distanceToHub.toFixed(3)}`)
-        })
-      }
-    }
-    
-    return hubPatterns
-  }
-
-  /**
-   * Create MSP traces for a hub pattern
-   */
-  _createMSPTracesForHub(hubPattern: {
-    hubTarget: string
-    hubPosition: { x: number, y: number }
-    components: Array<{
-      name: string
-      component: any
-      position: { x: number, y: number }
-      distanceToHub: number
-      pinName: string
-    }>
-  }) {
-    const { hubTarget, hubPosition, components } = hubPattern
-    
-    console.log(`🔗 Creating MSP traces for hub ${hubTarget} with ${components.length} components`)
-    
-    // Skip the closest component (it connects directly to the hub)
-    // Create chain connections for the rest: each connects to the previous in the sorted order
-    for (let i = 1; i < components.length; i++) {
-      const currentComponent = components[i]
-      const previousComponent = components[i - 1]
-      
-      const debugColor = this._getMSPDebugColor(i)
-      const debugStyle = this._getMSPDebugStyle(i)
-      
-      console.log(`🔗 ${currentComponent.name} (chain pos ${i}) → ${previousComponent.name} (chain pos ${i-1})`)
-      console.log(`🎨 ${currentComponent.name} debug trace style: color=${debugColor}, style=${debugStyle}`)
-      
-      // Create the MSP trace
-      const trace = this._createMSPTrace(
-        currentComponent,
-        previousComponent,
-        hubPosition,
-        i,
-        debugColor,
-        debugStyle
-      )
-      
-      // Mark the component so it doesn't create its own trace
-      currentComponent.component._mspRouted = true
-      
-      console.log(`✅ ${currentComponent.name} generated hub MSP trace to ${previousComponent.name}!`)
-    }
-    
-    // Mark the closest component (it will use normal routing to the hub)
-    if (components.length > 0) {
-      components[0].component._mspRouted = false // Allow normal routing to hub
-      console.log(`❌ ${components[0].name} uses normal routing (closest to hub)`)
-    }
-  }
-
-  /**
-   * Create an MSP trace between two components
-   */
-  _createMSPTrace(
-    fromComponent: any,
-    toComponent: any,
-    hubPosition: { x: number, y: number },
-    chainIndex: number,
-    debugColor: string,
-    debugStyle: string
-  ) {
-    const Trace = require('../Trace/Trace').Trace
-    
-    const fromSelector = `${fromComponent.name}.${fromComponent.pinName}`
-    const toSelector = `${toComponent.name}.${fromComponent.pinName}`
-    
-    // Calculate alignment Y based on hub position
-    const alignmentY = hubPosition.y - 1.25
-    
-    const trace = new Trace({
-      from: fromSelector,
-      to: toSelector,
-      schematicRouteHints: [
-        { x: toComponent.position.x, y: alignmentY }
-      ]
-    })
-    
-    // Set debug properties directly on the trace instance
-    trace._debugColor = debugColor
-    trace._debugLineStyle = debugStyle === 'dashed' ? 'dashed' : 'solid'
-    trace._debugWidthMultiplier = 1 + chainIndex * 0.3
-    trace._debugTransparency = 0.8 + chainIndex * 0.05
-    
-    // Add the trace to the subcircuit
-    this.add(trace)
-    
-    return trace
-  }
-
-  /**
-   * Get debug color for MSP trace based on chain index
-   */
-  _getMSPDebugColor(index: number): string {
-    const colors = ['#ff0000', '#00ff00', '#0000ff', '#ffff00', '#ff00ff', '#00ffff']
-    return colors[index % colors.length]
-  }
-
-  /**
-   * Get debug style for MSP trace based on chain index  
-   */
-  _getMSPDebugStyle(index: number): string {
-    return 'dashed'
-  }
-
-  /**
-   * Get component position by selector
-   */
-  _getComponentPosition(selector: string): { x: number, y: number } | null {
-    try {
-      const parts = selector.split('.')
-      if (parts.length !== 2) return null
-      
-      const [componentName, pinName] = parts
-      const component = this.selectOne(componentName)
-      if (!component) {
-        // Try different selector formats
-        const altComponent = this.selectOne(`.${componentName}`)
-        if (altComponent) {
-          return altComponent._getGlobalSchematicPositionBeforeLayout()
-        }
-        console.warn(`⚠️  Component not found: ${componentName} from selector ${selector}`)
-        return null
-      }
-      
-      return component._getGlobalSchematicPositionBeforeLayout()
-    } catch (error) {
-      console.warn(`⚠️  Error getting position for ${selector}:`, error)
-      return null
     }
   }
 }
